docs: eager load agent search context
