+++ docs/.env.example
@@ -12,6 +12,7 @@ INCEPTION_API_KEY=your_inception_api_key_here
 # INCEPTION_BASE_URL=https://api.inceptionlabs.ai/v1
 # DOCS_AGENT_MODEL_FLOW=gateway
 # DOCS_AGENT_GATEWAY_MODEL=openai/gpt-5.4-mini
+# DOCS_AGENT_EAGER_SEARCH=0 # disable automatic pre-model docs retrieval; enabled by default
 # DOCS_AGENT_SEARCH_PERF_LOG=1
 # DOCS_AGENT_SEARCH_LOG_QUERY=1 # includes raw user queries in server logs; avoid in production
